EditorElement Distinction
Working on splitting up the functional parts from the state and rendering parts. This is the first part of moving the state out of the class, and into it's own setup. Then the UI can just be rendered off of that itself. I'm only changing the name to help with my brain, and for during the migration.