Editor Prev/Next Methods
Working to move the methods off of the Editor state objects themselves, as they shouldn't be tied to the state itself, nor the element. This seemed like an unfortunate change with this kind of migration for me, I'm curious about the benefits that come with not using this, in the past I've seen this as a drawback, but I'm not sure if that's the case or not, I think I just haven't seen the alternative for it yet.