Focus On Open Fix
I think this seemed to break in the 'Editor Refs Shim' commit, or at least somewhere around there. It appears to be because I accidentally changed the order of when the editor is added to the global state, and then when opening that as a ref with `query()` it hadn't yet been added to the global state pool, so it couldn't find anything.

So moving the `open()` call up to `createEditor()` seems to fix it. The issue is that the animation was starting to run, before the element was attached to the state. I had to move the `focusedOverride` state onto the editor itself, since it has to be persisted from the constructor, to the now-delated, `open()` call.

Glad to have figured this out, I thought this was a recent bug I implemented, but this was actually from 4 days ago, and I hadn't even noticed! There are almost certainly going to be more occurrences like this, I think things are going thankfully well so far though.