Editor State + Ref Split Ups
It's still the same object under the hood as of this commit, I started migrating the types to where they use separate instances to handle things though. So it's the point of what handles what, which is what I'm trying to work on splitting apart.

If you paste this just at the end of the `query()` function, you can see it's still the same monster object being used for both.

```js
console.log(ref, editor, ref === editor);
```

Now, ideally, I should be able to change the rendering logic over to SolidJS, and just return the manual ref handling in the same way from `query()`.