Focus Override Instantiate Omit
This prop shouldn't be part of the create stage, it's only related to the opening handling, and it had to be present on the class only because of the moving from the `open()` call into `createEditor()` instead. So the instantiation shouldn't have anything to do with setting this, only when it opens it.