Card Instance Method Flattening
Working to make a purely functional component for cards! Realized I needed to do this step first, because it removes the need to have an instance to call it's methods publicly, now I can abstract how this works to the module itself, and not worry how the caller uses it.

This kind of thing is what makes TypeScript outstanding. This was relatively trivial to migrate over, and it would've been very hard to detect all of the edge cases of types here, and implement it properly.