Card Children Flattening
Abstracted away the internal workings of how Card components work, into now a singular definition. This will make things work much easier for inlining everything as JSX and migrating to signal-based state :)

I did just notice that the migration of Card methods over to general functions, it did seem to break the distinguishing of allowing Widgets to be closed while a Dialog is open. That was neat. I'll have to rework the backing for these and how they tell each other apart, in regards to the upstream app state changes.