CardItem Component
This is meant to help start abstracting further parts of the logic around Cards. I realized my paradigms were a big complex, and now that SolidJS is in the picture, more components are seeming to pop up in terms of things that can be narrowed into a single definition/implementation.