File System Access Check Fix
Not sure when this bug appeared, I tried to find it with Git blame on GitHub, but it didn't seem to come up multi-layers of commits into the search. I only just discovered it because of the linting that `tsc -w` provided. I'm not sure if it's the new TS version I'm using (probably that!!!), so thanks to that I'm fairly sure. I didn't have any errors when using it on the project last week.

Realizing aging is a very daunting thing.