Album Sort Learning + Editor Component Selections
Moved the editor component's selectors away from the `ste-editor` Web Component implementation, and to the now more generic `.Editor`. It was already using `.editor` solely for styling, but now I'm going to use that for the component down the road too (still just for styling ideally, but yeah, since the JS won't need selectors once everything is state/ref driven anyways).

Listened to some sadder songs for a bit, it did make me feel better and kind of reset again. Now I'm curious about what makes up a proper sad song for me, there's a certain kind of sad, not just the generic storytelling kind, like "she left me!!". More like the realization of your own fate, as well as the time you have left with others, and their fate too. That's just what my mind was spinning on, and that kind of took me for a loop. That kind of thing makes me want to even further have meaningful interactions with those I love and who mean everything to me. So 'existential' songs I guess.

Just did some final edits to Carpincho, I really am happy with the lineup of official songs for that one. It seems to be a solid EP of the three tracks, each very unbearable in their own right.

While talking with my dad yesterday at dinner, I think I realized the role of the producer/manager now, and I think I get how I can work on doing that part better for myself. It's a part of the puzzle for you as an artist to figure out whether a part of a song or a song itself is good in articulating your artistic expression, and or how that compares to the whole of the album or not. While sorting and final-mix checking songs today (for my two upcoming projects from this year), I realized there are the songs that I like because of what I meant in making them, then there are also the songs that I like for being a key piece to the full album. Now I realize, depending on the album's goals, you may only want to pick songs that work for both of those. One album may be off the wall, and all over the place, that's fine! If that's your intent, then just choose songs that you plain outright like. If the album has a certain aesthetic, then I would only choose the ones from that first set, which also work with the vision of the album. For my earlier projects (of the ones that I didn't go off the wall), I now realize that I chose all of the songs that I liked, regardless of whether they fit the vision I had for the album. I think that's why some of them didn't turn out as good as I'd liked. Now I see it's okay to set aside songs you like, if it isn't for the album. I (unknowingly) tried doing this new take with 'so, many, bad, decisions', and I am very happy with the conciseness of that one. You can see all of the bonus ones that didn't fit the vision over in the 'Colloquial Outtakes' sibling project. Some of those wouldn't have made the cut regardless, but others I have been 'partially bummed' that I didn't include on the main album, because I like the songs. Now I see that it was because it would dilute the album in conjunction with the songs that I thought were my 'top tier' lol. So now I see that method, and it makes a lot more sense.

If you're into music, great luck into sailing the seas of creativity! There's so much to learn, I love this journey and finding things out I've never thought of before. It really is wonderful.