EditorTab Component
Slowly breaking this one into it's own component, things are still breaking even though I'm trying for them not to. I think the issue is with the getters and setters for the refs not being updated by way of `query()` and the `EditorElement.state` property, which I'm trying to build the state off of.

https://docs.solidjs.com/reference/jsx-attributes/attr
https://github.com/solidjs/solid/issues/1101
https://github.com/solidjs/solid/issues/1689
https://www.solidjs.com/docs/latest/api#attr___
https://www.reddit.com/r/MacOS/comments/12xt6jt/do_you_use_natural_scrolling/
https://www.reddit.com/r/apple/comments/tkxtn0/universal_control_should_have_an_option_to_use/

This commit was from the last day or so, wanted to come back with fresh eyes the next day to ensure everything it was doing was correct. With the things I changed, I think it's mostly right. The only problem is the state-related parts I was mentioning before.