PreviewOption Component
I think this is working correctly, just want to make sure I didn't forget migrating anything. Awesome that things are working now!

Yeah so the setup does seem to be, you want to split up state from the elements themselves, and migrate things to JSX expressions. So maybe first you want to migrate children components over to JSX, keeping the current state. Then once everything is componentized-out, you can move it over to reactive state, since you already divied up where things are depended elsewhere. This was the main problem with this codebase, there was a BUNCH of crosstalk pollenation going on, where one component would modify another, because the app knows it's a singleton. That's not maintainable for scalability though, so a component shouldn't need to know about things other than what it needs to know about, if that makes sense. That's why state is so great, you separate the info out of the elements themselves.