Editor Sibling Refs Fix
Moving this over to solid references while things are currently setup. Solid's use of real values as components is very helpful for this. The issue is that I needed to access the elements before they were appended to the sibling components, and that's the way I had implemented the getter to work. Ideally down the road this will also be handled with Solid's state, rather than refs and adding/removing elements manually. Before having that though, I had to move all of these components over to proper JSX and contain their actual state.