Editor API Field Fallback Removal
Now the signals are used in favor of the instance-based getters and setters! Closer and closer!

Also discovered a bug in the migration of `getBasename()`, which it should have been imported into the Workspace module. The error hiding for the line that uses it obscured my discovery of finding the error.

Note that the necessary side-effects for `setName()` and `setValue()` aren't added yet with this one, I have to figure out that part, maybe I'll wait for the JSX migration, since that will be easy by then I think. Doing effects and state manipulation in this mixed state makes things harder to track, and I don't want to add a bugged implementation if I can just hold off with the broken feature for the meantime. Showing these fields as being unused does help make it easier to tell whether I've fixed this yet.

This codebase is so much better, slowly but surely getting closer to nice state management!!!