Editor State Flattening
Now things are using the dedicated signal state objects as where you set the state, rather than querying for the element, or getting things that way.

This brings things back to where the state of app-level parts are tracked with the state objects themselves, rather than querying for them with `query()`.