SCSS Config / Errors
Possibly thinking about moving away from SCSS/Sass directly, in favor of just plain CSS again. I don't use any of the advance features it provides (yet at least), and I was mostly only using it because I wanted to experiment with nesting, but polyfill it's usage in production so that could work nicely for anyone. Maybe I can give a new check to whether it's more supported cross-browser now anyways. We'll see.

https://github.com/sass/dart-sass/issues/2352#issuecomment-2358232959
https://vite.dev/config/shared-options.html#css-preprocessoroptions

I want to get this version of the project up and running in stable, it's been a long time since I made an update to the main site itself, and a lot of these reworks and changes will make adding actual new features something I can look into again. I've been wanting that for a long time now, and just unraveling the codebase has been a task to get that into something maintainable to where it makes sense to do.