Editor Autofocus Fix
This one was a deep dive to figure out what was going wrong, turns out it was the reactivity that was being applied in regards to updating state, and it should have only been applied to the signal value itself, rather than the effect-based state of the component. The ideal scenario is that a component shouldn't care when it's state is updated though, so in theory this shouldn't be an issue, and it would be able to account for it properly. This setup is still very fragmented though, so the timings are reactivity is a bit mixed up.

https://github.com/Offroaders123/Smart-Text-Editor/blob/bd1b8e412b542871e881c606a0121ece5dac456c/src/Editor.ts#L132

Re-focusing after exiting a dialog still isn't working, have to debug that.