Dialog Close Focus Fix
The issue was that when closing a dialog, the previous focus state was negated because the dialog check was wrong, so when creating a new editor with the dialog open, it would actually focus that editor, behind the dialog. This was actually just a bug in the logic, from what it looks like. I'm trying to think if I just wrote it wrong, or if I'm missing a case here, and the previous one was intentional. Might go back into the git history to see where it might have flipped unsuspectingly.

Bullys (sic)
My Dilemma

Forgot to include this link in the previous fix.
https://github.com/Offroaders123/Smart-Text-Editor/blob/main/src/Card.ts#L164