Revert Focus Debugging
Reverting the debugging from the previous few commits, now that the fixes themselves have been discovered.

I love "Blameless (The Floating Face)"

I think I realized I really like the aesthetic of Boil That because it has hints of grunge in it, mixed with Keneally's unique style, and both of which are some of my favorites.