Automatic Refresh Fix
Turns out it was just a logic/boolean-null coalescing issue. The error wasn't raised in TS because I was using `!` at the start of the property check, which was let alone flipping the state to be incorrect, but it was also removing the nullish part of that, hence hiding the broken logic.