V5! 🚀 ❤️‍🩹
Wow, this is spectacular!

With just over 180 commits, and slow/steady work over the last 14 months, I moved and refactored the project fully over to SolidJS, leaving things exactly as they were, near exactly.

This has been a very complex process, and it took lots of learning and things to figure out to get it working.

I fixed a bunch of bugs that fell through the cracks over the years too, now settings works correctly again, and separated everything into it's own component and state. This was the most complex part of it all, getting the app to run the exact same way, while essentially reinstalling and rebuilding the entire wiring harness itself. I know most people would rather build a new app from scratch from the ground up, but I really wanted to do this to learn *why* things had become unmaintainable, and figure out how to leave that quicksand, and get out into the light again. I still plan on doing a very nice rework for the real usaable parts of the app, I wanted to migrate all of the existing logic over to more explicit concepts though, like state, signals, effects, modules, strictly-typed components, everything great.

Lifesigns - At the End of the World

The first commit of this long process (to which this migration turned into eventually) was when I moved over to SCSS, to get nesting working (with polyfilling).

It's crazy that STE is about to turn 5 as well, it was winter 2019, just before the pandemic when I started building it. Now this project is in the best state it's been in, and much more prepared for a proper redesign (don't worry, I love the current one, I want to try new things though).

I also removed the macOS-specific styling, as it looked really cool, but now I feel that the future of STE is to have a unified style across all platforms, it might feel more unique and at home, no matter where you use it.

I think this year was a very interesting one for my learning further towards coding, it's okay to not commit straight to stable super fast, big changes take time, and making sure you try your best at them is more important than getting them done, if you want a solid maintainable product. There was so much more work done here than I remembered doing, it's crazy! I have felt like this project has fallen behind the last two years or so, and now it's much more prepared for allowing new features to be added, and that's been what I've wanted to start looking into again for a long time now. I didn't want to do that on top of the technical debt I was accumulating though, so it felt like that had to come first. I'm so glad I did that.

Cheers to long-running projects, life is awesome 🥂