Scrollbar Styling Fix
Figured out why the scrollbar styles were all wack in Chrome, it's because of when it added support for `scrollbar-color`, which is a good thing! I think I'll properly migrate to that solelyby itself, down the road, maybe with the redesign.

https://frontendmasters.com/blog/heads-up-on-custom-scrollbars-chrome-is-supporting-the-standard-now-which-overrides-the-old-pseudo-elements/
https://developer.chrome.com/docs/css-ui/scrollbar-styling

You can tell I'm not as much in touch with solely frontend work, this was in January of this year 😭😆

```ts
/**
 * Creates a new Editor within the Workspace.
*/
export function createEditor(options: EditorOptions = {}): void {
  const editorElement = EditorElement(options);
  const { state } = editorElement;
  setEditors(state.identifier, state);

  const autoCreated = state.getAutoCreated();
  const focusedOverride = state.getFocusedOverride();
  if (state.isOpen || activeEditor() === null){
    open(state, { autoCreated, focusedOverride });
    if (autoCreated) {
      setView("split");
      // state.ref.value = "xcvxcvxc";
      insertTemplate("html");
    }
    state.setFocusedOverride(false);
  }
}
```