STE App State
Now I see how the main app state can now coincide with that of the root component, and it makes sense for that to contain everything, and it can pass all of the props to the child components as needed. A component should only have access to explicitly what it requests to be able to change and or apply or define. A better way to put it, it's like if a component only needs to be able to read a file's name, it doesn't need to write it's name or know it's contents. Things like that.

So right now I only just moved the state into the same module as the main component, now I have to actually move it into the component itself, which will allow for more explicit state definitions, in regards to what a component needs to be able to run and calculate, and adjust things correctly.

Next to the light, the sun shines bright, piña colladas in the rain; wrath