Card State Creep
I think the issue with STE's state management, which I now understand a bit better, is that it shoves all of the component-specific side-effect code into a single handler, and this is intrinsically tied to the function that updates the state itself too. So let alone is the side-effect code not independent of the state, but it is also implemented outside the scope of a component context, meaning it has to be declared and managed globally. That's why much of this codebase is so crazy to get fixed up.

So now I see that the side-effect code is best declared inside the component itself, and the state only has to worry about it's value and when it changes. It will then send this update down the signal tree, depending on who's listening downstream. So a top-level signal doesn't explicitly have to know how it will effect children that rely on it, that is for them to figure out on their own.

So I'm going to keep this commit for sake of learning. Now I get where the line works easiest to be drawn with this. The issue isn't the passing of state into these functions, it's that the functions shouldn't be singleton-based in the app like that, they should be individual handlers for each of the components' state.

Fender Bender Ascender