Nullish Effect Ref Usage
I think these appear to be fine, to me at least. I might be missing something here, and maybe I had kept these type errors for some reason, to imply something else that had to be worked on. I think we're good though? The only thing I can think of might be the Web Component lifecycle for Menu Drop, which might be throwing things off. I'm not so sure.