Toggle Bring-Back
Another Look at Space Animals

Realized my commit from a few months ago unintentionally removed the ability to toggle the cards by way of shortcuts and app menus. It's because of the change in state models.

I love the sound of the glitchy part in this song. I very much want to expand on it again later in new projects I make in the future.