Card Active State Setters
Okay, almost moving away from the effect-based state management that STE has been using since the dawn of time. So so so freaking close!!! It's freaking insane how twisted an obtuse the codebase was. It's okay though, this was my first real project essentially.

Absolutely check out Aspid Extravasation. Been the new album of mine for the past week or so I think, absolutely outstanding. Unbelievable.