openCard() Effect
This is in progress. I am translating parts of these methods into effects based onto the state they are derived from. This is conceptually really complex for me, I still have to get my mind around everything it's doing here. I think I might be either making it way more complex than it seems to be, it's something I'm not totally understanding of yet, or maybe it really is just that complex and messy at the same time.

The audio for this video sounds just so great on my headphones, I love it. It's a very nice contrast to that of the complex music I listen to throughout the day. I love both of them just as much. This counterpart is relatively new in my media enjoyment, as of the last few years. Mindfulness? Meditation?
https://www.youtube.com/watch?v=k4Egjey8oWE