Dialog Closing on Widget Open
Trying to figure out the original logic I wrote, and make it effect-based now. A lot of the if guards are making it very confusing to figure out the intent of what it is meant to do, and what feature it is meant to toggle/open/close. The whole function stack is just a mishmash of concerns. Effect code linked to state, with an inheritance model of things which should be defined separately on their own instead. Let alone, everything could be using Promises too. And of course, signals. Those would have made much more sense. I knew no better though, so that's okay 👀

Everyday Ghosts