Card Activity Reactivity
Aah, now I see what the issue is! Same with this version too.

So I just have to hook these two signals together. They are essentially a 'memoized' getter and setter, if that makes sense. One can change, and that will change the other. The other will change, and it will update the first one. Is this dual-directional binding? Something like that. I think Solid can do it, React can't. We'll see.