Card Abstraction Flattening
Going to try to reverse that which is the dismay of my pain with this project: The Card System. For some reason I made these into a single abstraction, and it is still plaguing me 5 years later. Wow.

Just going to rebuild them on their own now, at this point. Even if it changes the functionality, just slightly. The codebase can't keep hanging on to this old cruft anymore, it's really holding it back. The inheritance here really destroyed things.

I think a big part of the issue I made was that I used inheritance, without actually making subclasses. So it was just a big ugly single class, build on more than one child implementation. Very eew.

Marco Minnemann Orchids, it's a great album!

Not to be particular, I think I sonically like the mixes/mastering on Marco's older work. It might be because he does the mixes on his own at home? I think the originals might have been in a studio, but I'm not totally sure. With my new ear for that kind of thing, I can pick up now on why his new projects haven't quite clicked in the last year or two. I think I started to notice it before I was able to with the original ones I found.