Widget Impl
Okay, this seems to be working!

Listening back to my past mixes now to figure out where I might need to work on things. So far it seems like the main issue is just automation, and EQ. Both of which have been in need of help for a bit now.