Alert Impl
Wow, can't believe just how simple this really is.

If anything, now I see that it makes sense to just make sub-components for this, and reuse those composably, rather than make an inheritance model for these.

Composition wins!

I think Limerence might be my best song yet. I know that's a big statement, but it really checks all of my boxes with my goals right now. The mix is great, it's creatively free, and it has a lot of depth.

Really looking forward to where I take things from here. Having the dedicated mic for things like that will only make it more precise and personal.

Also realizing, I am discovering a new way that I will work on albums. I don't have to have a single way to do it. That's the magic with it.

From the projects through this year so far, I plan to pick things from the sessions, and only pick the like-songs and ones that make sense in an album context. It doesn't have to align to the time periods or anything. While writing these, I think in the back of my mind I had consciously decided it was okay for each of the songs to be used in contexts other than the ones they were recorded in. I am really hoping to drive that home this time around, once I get to these songs sessions. We'll see though. It can only be for working towards the goal of benefiting the music listening experience, or the goals of the music itself.