Widget Minimize State
I think I realize that this doesn't need to be global? I'll have to double-check that though. It might make sense to just have this be based on each Widget individually, and have it be effect-driven, based on when that instance minimizes, and closes. That's the original behavior I wanted to have early on, if I remember right.

This is now the cusp of my abilities to design great interfaces. I know how I wanted the interactions to work, from very early on. I did not have the programming experience to design things properly around state and effects though, and that's what was holding me back. I can make great things now, and I can fully see that! Very happy about this! So glad I started looking into functional frameworks.

Coming up on 6 years I've been programming fully, now. Outstanding. Been playing guitar for about 10 years too. Same with recording in GarageBand to some extent. Moreso making music for about 8 or 9 though. Consistently for the last 4 just about. Wow, that also means I've been working for the last 4 years as well. That's outstanding.

Be happy for yourself, you have lots to be proud for! 😃⭐️