Minimize Handler
Okay I do need it in global state. I need some way to be able to update the signals, since the global keyboard shortcuts need access to it.

I think the pattern here is to use dependency injection, with a callback like this. Then it allows other scopes to call it without needing to know the scope that appended it. This is yet another paradigm I didn't see, nor understand early on. This allows you to change the code, without other code knowing how it was changed. It simply only needs to check whether it is something callable or not. This is very easy to manage in TS as well.