Minimize Handler Signal
Okay, this is the answer! Instead, I can use a signal of signals. This seems to work great. The only thing broken is that if you open a Widget while another is open (replacing it, this is expected behavior), there is a race condition in that the setter will be set to null after it was already properly defined to the newly opened Widget.

I can also get rid of my setMinimizeHandler() function, since this is merely just a wrapper for the signal at this point. I wonder if this is what is causing the race condition.

Guthrie Govan: Fives, Uncle Skunk, Sevens, Eric. This album is great.