Handler Alias Removal
Just wanted to do this separate.

Was going to leave this in for later formatting, but realized that's not being proactive. Don't let your codebases rot at your own will!

Listening to Elkosaurus Tex, I do love this album. It holds a special place in my heart. Trying to get my brain back to where I originally wrote it, at an understanding level. It was pure muse and admiration of the things I was making fun and light of.

Realized something earlier. I am okay with my music sounding repetitive across new and old releases, because no matter how I look at it, things are never coming from the exact same place. They can't. I change over time, and there are always new perspectives involved with the new work. It might be from a nostalgic point of view, or a way of looking back on that original feeling, or even just bringing it back to the forefront again. But even with that lens, it is through different eyes that time around. It's never quite exactly the same, in *the* first-person point of view. It's always from *a* first person point of view that it's driven from, but it's not the same *person*. Maybe parts are similar, but there is so much change. It's drastic.

It's not that I want to go back, or be exactly like that time period again. It's that I want to learn how to reencorporate that experience of me again, back to the forefront, to keep it alive in the new set of eyes that I have. That's what I'm striving for.