Full Unset Minimize (broken)
This is running too often. I think it's because it's running for each widget.

Maybe this should be a global effect rather than a local one? That might not make sense in terms of the component though, in being able to use the component without worrying about where it's placed.

So, this isn't quite working now, but I think I know why.