Minimize Change Reset Pause
Going to come back to this once it is clearer. It doesn't seem to need to be added yet, the animation is working at least. This will be necessary though for things that need the global state hook though.

Aaah, it might actually be a two-way binding issue again, come to think of it.