Menu Drop Touch Fix
It still doesn't seem to have fixed it, which is really annoying. Going to debug this. The test file on that repo, it is working correctly. In this app setting, it isn't. What the heck??

Edit: Just as typing this, I figured it out. It's because it is still being loaded with JSDelivr. I swear, this project is still a mess.

https://github.com/Offroaders123/Menu-Drop/releases/tag/1.1.0