version 5.8.0
Okay, catching up the main branch now to what I've worked on the last few months. The main fix is that MenuDrop stopped working on iOS, and I've been meaning to fix that for a while. Now this update addresses that. Yay! This took a bit too long, I was getting frustrated when working on it though, so I had to leave it for later (which seemingly appeared to be now, for whatever reason).