GitHub Actions Update
Guess it's been a while since STE was updated, I think I remember meaning to do this a while back too, when I was working on my https://offroaders123.github.io/ page. I had to fix this same issue then, but I hadn't brought it over to STE or Dovetail yet.

Running the action on the previous push failed with this link as a reference.

https://github.blog/changelog/2024-04-16-deprecation-notice-v3-of-the-artifact-actions/