App Root State Signals
Moving the big state handles over from DOM attributes to proper signals.

This is going to have to transition incrementally. Things to some extent are still inherently tied to the DOM representation itself. I didn't know there was an alternative to this back when I started this project, and I don't know if I would have understood why you would separate the two as well. I didn't understand "state" for a long time, back then. Now it just plain makes a lot of sense. Learning is so cool, you really start to counter yourself the more examples you see, haha. Proud of how far I've come to be able to see the distinction of things like this now! It's even cooler that this project is still alive and kicking. If anything, I have hopes that it will be much better than it currently stands, in new light once again. This project can be really strong, it just needs a great foundation to stand on.