View & Orientation Properties Removal
Now they are simply states, rather than attributes which hold the current state. Heck yeah.

Note that the classes are still applied, these are necessary for styling. This should be moved to effects though, and later down the road, just inlined effects in the JSX itself, once we get to that point.

Part of how I know migrations like these are safe, I search the whole codebase for usage of things like `data-view` and `data-orientation`. I check how it needs to be replaced, whether it is still needed, and if I can just remove it.

I think I used to store these as attributes because I used to use it as styling as well as state. So you would only apply the attribute, and it would update the styling too. A nice hack. If relying on the DOM wasn't an issue, this is a great way to build things simply. It is very declarative, come to think of it. You can get really reactive code working nicely with a vanilla web stack, it just is not very declarative if you want to remove the DOM counterpart itself.