View Change Signal (& big music project details!)
Sit Back, Relax
The Rain is Back But Not Inside

I think I might start working on my projects from earlier this year, and the end of last year. Left off with around August's releases if I remember right. Almost all caught up!

I had some big ideas that I thought of when falling asleep last night, got up to go write them in my phone probably about 20 times lol. It's a big group of projects, which consist of being solely single-genre. That's something else new I have never done before. Usually I am trying to make something totally new, and it ends up mixing multiple genres together. It is hopefully even going to be ones that I don't have much interest in listening to on my own, that's why it's so intriguing to me. It's the point of learning about different genres I think, even if I don't have to necessarily "like" them, before or after. The point is that I am going to make the best of that style that I can, and it will be under a unique artist and album name, almost like a facade band. Unintentionally similar, it kind of reminds me of the early intent of Porcupine Tree, how it was a mythical 70s prog band which didn't actually exist, and Steven Wilson did all of the work for it. I am not trying to do exactly the same thing in practice, but in appearance, it is something very similar. For me with this project, it is more about trying to get "the aesthetic", even if it's cheesy, based in generalizations, or tropes. That's kind of the point of what it will be. It will be my solid attempt at that thing. Not for the sake of it being something I am paying homage too, it's different. More like a way of trying to enact empathy to that style, and or where it comes from. I want to understand what makes that thing click, and see if I can make it click for me too. I think part of this stemmed from realizing that I "don't yet understand how to enjoy" flat mixes, like modern metal productions. It seems much more perfected, but that's not what my ears look for. In realizing that I look for dynamics, and a live-feel, when a production misses that, it simply just doesn't hit home for me. So I wanted to learn why that's not the case for others. What does make it click for them? Is it something I can learn too? There is plenty of music out there in that style, and it seems like I'd be doing myself a disservice in not trying to understand what makes it work. There's always something that's a takeaway and or learning experience to be had.

Say for example, for my big project of albums and styles, I will have to learn how to utilize flat production and or straightforwardness in writing styles, as well as writing sections. I am not doing so simply just to be more accessible. I still have the thought that, if my music was liked by someone, that's a cool thing. I am not aiming for that though, like that being the only goal. My music currently stands as what I use to learn about things. It's kind of like how I learned how to code making a text editor, it's a recursive learning pattern. That's the same thing here. I had this tendency to assume that making things straightforward was simply for the goal of making things more accessible because you wanted to make money, be famous, and or get an audience/following because of your ego. That is my bias/assumption to certain styles. I now see that that's faulty thinking. I am learning these styles because they allow me to express things that I wouldn't otherwise be able to express with the palette that I currently provide myself.

But for right now, this work won't inherently/intentionally be related to or about my trip. If it ends up that way unintentionally, that's fine. But I don't have the goal of accurately representing myself with these new lenses, at least not yet. I would want to get used to these techniques before trying to use them to articulate myself.

Yeah, so it will be grindcore/thrash, pop (yeah, like Kesha, Katy Perry), 70s stuff like McDonald, rap, a bunch of things.

It also will come with a perceived "persona" that I made up for these "bands". Again, they aren't meant to represent myself, but more to break down the tropes of how I see the genre. It isn't necessarily meant to be tongue in cheek too though, either. It isn't meant to make fun of it, in the complete sense. It might be very unoriginal though.

I think part of this might be driven of me feeling like I've expressed a lot of what I've wanted to express for a while, and I need new tools to express myself creatively.

It might be the "pause" for my solo work, for the first time ever too. I may release my own work at the same time, may not. It's intentionally up in the air to find out what happens.

I am very curious about what happens when I try to be intentional about things too, because usually my styles drift between a bunch of different things. If it is meant to intentionally be straightforward, I am very curious about what that will look like. I have the power to make things in that regard, but sometimes now the self control to omit the things I love from allowing the record to be straightforward.

Take the Sunshine
Browniedistance